refactor(web): share one graph renderer for plan + state graphs; module-organize state (#765)
The state graph and plan/impact graph are now the same graph — extract a shared
<ResourceGraph3D> with the module-cluster force, translucent per-module boxes,
module-stripped labels, centroid framing, and click-to-highlight blast radius.
impact-graph and state-graph-3d become thin wrappers differing only in what
colour encodes (plan: change action; state: a chosen pivot — resource type /
module / provider / managed-vs-data), so they can't drift apart.

- State graph now organizes by module identically to the plan graph (server
  emits module='' for root, matching the impact convention so the cluster
  force skips root).
- Overlay panels no longer crowd the canvas: a small always-on toolbar (Reset
  view / Clear / Key / Resources); Key (legend) + Resources (list) are opt-in,
  dismissable overlays, off by default — mobile-friendly.
- Relabel the pivot 'Color by' (clustering does the spatial grouping now).

Verified live on Tilt: 3-module state clusters into module.network /
module.compute / module.database boxes; blast radius + version picker work.
