+++ docs/state-resource-graph.md
@@ -8,22 +8,33 @@ view: estate shows how workspaces relate to each other; this shows how the
 resources within one workspace relate. It answers *what hangs off this VPC? what
 would ripple if I touched this security group? what's the shape of this state?*
 
+It is the **same interactive graph as the plan/impact graph** ([impact-graph.md](impact-graph.md)) — same module clustering, same click-to-highlight blast radius — sharing one renderer. The only difference is what **colour** encodes: the plan graph spends colour on the change action (create/update/delete); the state graph has no "what's happening" axis, so colour is free to encode a pivot you choose.
+
 ## What it shows
 
 - **Nodes** — every resource in the state, one per resource address (a resource
-  with `count`/`for_each` collapses to a single node). Managed resources render
-  as spheres, **data sources** as boxes. Node size reflects in-degree: the more
-  resources depend on a node, the bigger it is, so foundational resources stand
-  out.
+  with `count`/`for_each` collapses to a single node). Node size reflects
+  in-degree: the more resources depend on a node, the bigger it is, so
+  foundational resources (a VPC, a shared security group) stand out.
+- **Module clustering** — resources are grouped spatially by their module and
+  drawn inside a translucent `module.<name>` box; root-module resources float
+  free. So a multi-module state reads as distinct regions at a glance.
 - **Edges** — `depends-on` relationships Terraform records in state (both
   explicit `depends_on` and the implicit references it tracks per resource).
-- **Group by, your choice** — colour the graph by **Resource type** (the
+- **Blast radius** — click a resource (or pick it from the **Resources** panel)
+  and everything that transitively **depends on it** lights up, so you can see
+  what a change would ripple into.
+- **Color by, your choice** — colour the nodes by **Resource type** (the
   default — every `aws_subnet` the same colour), **Module**, **Provider**, or
-  **Managed / data**.
+  **Managed / data**. (Colouring is independent of the module clustering above.)
 - **State version picker** — defaults to the workspace's **current** state
   version; drop the picker back to **any older version** to graph a previous
   state (Terrapod versions every state upload, so history is available). Great
   for seeing how the resource graph changed across applies.
+
+A compact toolbar over the graph keeps `Reset view` and `Clear` always to hand,
+with `Key` (the colour legend) and `Resources` (the searchable resource list)
+as dismissable overlays so the graph keeps the canvas.
 
 ## Graph and table
