+++ web/src/components/state-graph-3d.tsx
@@ -1,188 +1,75 @@
 'use client'
 
-// State resource graph 3D renderer (#765). Client-only (WebGL): the state-graph
-// tab imports this via next/dynamic { ssr: false }. Pure renderer — the tab
-// owns the fetch and passes the graph in, so the same data also drives the
-// accessible table fallback. Managed resources render as spheres, data sources
-// as boxes; both are coloured by the chosen grouping axis.
-import { useEffect, useMemo, useRef, useState, type ReactElement } from 'react'
-import ForceGraph3D from 'react-force-graph-3d'
-import SpriteText from 'three-spritetext'
-import * as THREE from 'three'
+// State resource graph (#765) — a thin wrapper over the shared <ResourceGraph3D>.
+// Same graph as the plan/impact graph (module-clustered resources, dependency
+// arrows, click-to-highlight blast radius); the only difference is what colour
+// encodes — here it's the user-chosen pivot (resource type / module / provider /
+// managed-vs-data), since state has no "what's happening" axis to spend colour
+// on. Client-only (WebGL): the tab imports this via next/dynamic { ssr: false }.
+import { useMemo } from 'react'
 import {
   categoryOf,
-  endId,
   PALETTE,
-  EDGE_COLOR,
-  type StateEdge,
   type StateGraphData,
-  type StateNode as BaseNode,
+  type StateNode,
 } from '@/lib/state-graph'
-
-// The renderer stashes mesh/sprite refs on nodes for live recolouring.
-type StateNode = BaseNode & { __mesh?: THREE.Mesh; __sprite?: SpriteText }
-
-// Minimal imperative API surface we use (published types don't survive dynamic).
-interface D3Force {
-  strength?: (n: number) => D3Force
-  distance?: (n: number) => D3Force
-  distanceMax?: (n: number) => D3Force
-}
-interface FgMethods {
-  zoomToFit: (ms?: number, px?: number) => void
-  d3Force: (name: string, force?: unknown) => D3Force | undefined
-}
-type FgProps = {
-  ref?: React.Ref<FgMethods>
-  width?: number
-  height?: number
-  backgroundColor?: string
-  graphData: { nodes: StateNode[]; links: StateEdge[] }
-  cooldownTicks?: number
-  onEngineStop?: () => void
-  nodeThreeObject?: (n: StateNode) => object
-  onNodeClick?: (n: StateNode) => void
-  linkColor?: (l: StateEdge) => string
-  linkWidth?: (l: StateEdge) => number
-  linkDirectionalArrowLength?: number
-  linkDirectionalArrowRelPos?: number
-  linkDirectionalArrowColor?: (l: StateEdge) => string
-}
-const FG3D = ForceGraph3D as unknown as (props: FgProps) => ReactElement
+import { ResourceGraph3D } from '@/components/resource-graph-3d'
 
 export function StateGraph3D({
   graph,
   groupBy,
-  onSelect,
-  selectedId,
+  subtitle,
 }: {
   graph: StateGraphData
   groupBy: string
-  onSelect: (n: StateNode | null) => void
-  selectedId: string | null
+  subtitle: string
 }) {
-  const fgRef = useRef<FgMethods | null>(null)
-  const wrapRef = useRef<HTMLDivElement>(null)
-  const [size, setSize] = useState({ w: 800, h: 600 })
-
-  const data = useMemo(
-    () => ({ nodes: graph.nodes, links: graph.edges.map((e) => ({ ...e })) }),
-    [graph],
+  const categories = useMemo(
+    () => [...new Set(graph.nodes.map((n) => categoryOf(n, groupBy)))].sort(),
+    [graph, groupBy],
   )
-  const adj = useMemo(() => {
-    const a: Record<string, Set<string>> = {}
-    graph.nodes.forEach((n) => (a[n.id] = new Set()))
-    graph.edges.forEach((e) => {
-      a[endId(e.source)]?.add(endId(e.target))
-      a[endId(e.target)]?.add(endId(e.source))
-    })
-    return a
-  }, [graph])
+  const colorOf = (n: StateNode) =>
+    PALETTE[categories.indexOf(categoryOf(n, groupBy)) % PALETTE.length]
 
-  const colorMap = useMemo(() => {
-    const cats = [...new Set(graph.nodes.map((n) => categoryOf(n, groupBy)))].sort()
-    const m: Record<string, string> = {}
-    cats.forEach((c, i) => (m[c] = PALETTE[i % PALETTE.length]))
-    return m
-  }, [graph, groupBy])
-
-  useEffect(() => {
-    if (!wrapRef.current) return
-    const el = wrapRef.current
-    const ro = new ResizeObserver(() => setSize({ w: el.clientWidth, h: el.clientHeight }))
-    ro.observe(el)
-    setSize({ w: el.clientWidth, h: el.clientHeight })
-    return () => ro.disconnect()
-  }, [graph])
-
-  const near = selectedId ? adj[selectedId] : null
-  const baseColor = (n: StateNode) => colorMap[categoryOf(n, groupBy)] || '#64748b'
-
-  // recolour meshes + labels when pivot or selection changes (no full rebuild)
-  useEffect(() => {
-    const dim = (id: string) => selectedId && !(id === selectedId || near?.has(id))
-    for (const n of graph.nodes as StateNode[]) {
-      if (!n.__mesh) continue
-      const c = new THREE.Color(baseColor(n))
-      if (dim(n.id)) c.lerp(new THREE.Color('#0a0e17'), 0.72)
-      ;(n.__mesh.material as THREE.MeshLambertMaterial).color = c
-      if (n.__sprite) n.__sprite.color = dim(n.id) ? 'rgba(150,160,180,.4)' : '#e2e8f0'
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [groupBy, selectedId, colorMap, graph])
-
-  function nodeObj(n: StateNode) {
-    const grp = new THREE.Group()
-    const s = Math.max(3, n.indeg * 1.1 + 3)
-    // data sources are boxes, managed resources spheres — a shape cue on top of
-    // the grouping colour.
-    const geo = n.mode === 'data'
-      ? new THREE.BoxGeometry(s * 1.5, s * 1.5, s * 1.5)
-      : new THREE.SphereGeometry(s, 16, 16)
-    const mesh = new THREE.Mesh(geo, new THREE.MeshLambertMaterial({ color: baseColor(n) }))
-    n.__mesh = mesh
-    grp.add(mesh)
-    // label with the address minus any module prefix (kept short); the sprite
-    // still identifies the resource without the noisy module path.
-    const short = n.name.replace(/^module\..*?\.(?=[^.]+\.[^.]+$)/, '')
-    const sprite = new SpriteText(short)
-    sprite.color = '#e2e8f0'
-    sprite.textHeight = 3.4
-    sprite.backgroundColor = 'rgba(10,14,23,.6)'
-    sprite.padding = 1.5
-    ;(sprite as unknown as { material: { depthTest: boolean } }).material.depthTest = false
-    ;(sprite as unknown as { position: { set: (x: number, y: number, z: number) => void } }).position.set(0, s + 5, 0)
-    n.__sprite = sprite
-    grp.add(sprite)
-    return grp
-  }
-
-  function frame() {
-    fgRef.current?.zoomToFit(600, 90)
-  }
-
-  useEffect(() => {
-    const fg = fgRef.current
-    if (!fg) return
-    fg.d3Force('charge')?.strength?.(-130)?.distanceMax?.(220)
-    fg.d3Force('link')?.distance?.(52)
-    const nodes = graph.nodes
-    const centre = (a: number) => {
-      for (const n of nodes) {
-        if (n.x == null) continue
-        n.vx = (n.vx || 0) - n.x * 0.05 * a
-        n.vy = (n.vy || 0) - (n.y || 0) * 0.05 * a
-        n.vz = (n.vz || 0) - (n.z || 0) * 0.05 * a
-      }
-    }
-    ;(centre as unknown as { initialize: () => void }).initialize = () => {}
-    fg.d3Force('centre', centre)
-  }, [graph])
-
-  const linkLit = (l: StateEdge): boolean => {
-    if (!selectedId) return true
-    const s = endId(l.source), t = endId(l.target)
-    return (near?.has(s) || s === selectedId) && (near?.has(t) || t === selectedId)
-  }
+  const legend = categories.slice(0, 18).map((c) => (
+    <span
+      key={c}
+      className="flex items-center gap-1.5 text-[11px] font-semibold px-2 py-1 rounded-full bg-slate-700/25"
+    >
+      <span className="w-2 h-2 rounded-full" style={{ background: PALETTE[categories.indexOf(c) % PALETTE.length] }} />
+      {c}
+    </span>
+  ))
 
   return (
-    <div ref={wrapRef} className="absolute inset-0">
-      <FG3D
-        ref={fgRef}
-        width={size.w}
-        height={size.h}
-        backgroundColor="#0a0e17"
-        graphData={data}
-        cooldownTicks={180}
-        onEngineStop={frame}
-        nodeThreeObject={nodeObj}
-        onNodeClick={(n) => onSelect(n.id === selectedId ? null : n)}
-        linkColor={(l) => (linkLit(l) ? EDGE_COLOR : 'rgba(120,130,150,.1)')}
-        linkWidth={() => 1.2}
-        linkDirectionalArrowLength={3.2}
-        linkDirectionalArrowRelPos={1}
-        linkDirectionalArrowColor={() => EDGE_COLOR}
-      />
-    </div>
+    <ResourceGraph3D<StateNode>
+      nodes={graph.nodes}
+      edges={graph.edges}
+      title="State graph"
+      subtitle={subtitle}
+      legend={legend}
+      hint={
+        <>
+          Each sphere is a resource; each arrow points to what it depends on. Click a node → the
+          resources that transitively <b>depend on it</b> light up.
+        </>
+      }
+      colorOf={colorOf}
+      nodeSize={(n) => 2.5 + Math.min(n.indeg, 10) * 0.35}
+      sortNodes={(a, b) => b.indeg - a.indeg || a.id.localeCompare(b.id)}
+      renderDetail={(n, downstream) => (
+        <>
+          <div className="font-mono text-xs text-slate-100 break-all">{n.id}</div>
+          <div className="text-[11px] text-slate-400 mt-1">
+            {n.mode === 'data' ? 'data source' : 'managed'}
+            {n.provider ? ` · ${n.provider}` : ''}
+            {n.module ? ` · ${n.module}` : ''}
+          </div>
+          <div className="text-2xl font-bold mt-1.5">
+            {downstream} <span className="text-xs font-medium text-slate-400">depend on this</span>
+          </div>
+        </>
+      )}
+    />
   )
 }
